Modify hotel methods

+++ traveleasy-frontend/src/app/my-booking/my-booking.component.ts
@@ -31,6 +31,7 @@ export class MyBookingComponent implements OnInit, OnDestroy {
   showAsc: boolean = false;
   selectedRange: Filter = this.ranges[0];
   bookingSubscriptions: Subscription[] = [];
+  errorString: string = '';
 
   constructor(private bookingService: MyBookingService){}
 
@@ -44,15 +45,28 @@ export class MyBookingComponent implements OnInit, OnDestroy {
     }
   }
 
-  deleteBook(id: number): void {
-    const secSubscription = this.bookingService.deleteBooking(id).subscribe( data => {
-      console.log('Delete booking', id);
-      const thirdSubscritipn = this.bookingService.getBookings('').subscribe(data => {
-        this.bookings = data;
-        });
-        this.bookingSubscriptions.push(thirdSubscritipn)
-    });
-    this.bookingSubscriptions.push(secSubscription)
+  deleteBook(id: number, startDate: string): void {
+    const currentDate = new Date();
+    const bookingDate = new Date(startDate);
+    const currentYear = currentDate.getFullYear();
+    const currentMonth = currentDate.getMonth();
+    const currentDay = currentDate.getDate();
+    const currentWithoutTime = new Date(currentYear, currentMonth, currentDay);
+    const bookingYear = bookingDate.getFullYear();
+    const bookingMonth = bookingDate.getMonth();
+    const bookingDay = bookingDate.getDate();
+    const bookingWithoutTime = new Date(bookingYear, bookingMonth, bookingDay);
+  if (bookingWithoutTime > currentWithoutTime){
+      this.errorString = '';
+      const secSubscription = this.bookingService.deleteBooking(id).subscribe( data => {
+        const thirdSubscritipn = this.bookingService.getBookings('').subscribe(data => {
+          this.bookings = data;
+          });
+          this.bookingSubscriptions.push(thirdSubscritipn)
+      });
+      this.bookingSubscriptions.push(secSubscription)
+    } else
+      this.errorString = 'Cannot delete booking.\n The start date is in the past.';
   }
 
   applyRangeAndSort(range: string, sort: boolean){
